Feat(Animations): Added animations and hopefully prevented car yeeting.
Added some additional documentation and default mouse pointer but it needs to be scaled to screen resolution eventually so they all start on a certain one.

Car may still flip around depending on how the phone prop is created. but I tried to midigate it the best I could without adding a citizen wait to phone creation.