RFR: Renamed file, adjusted manifest, and refactored code.
Renamed the event so its more clear what it does and added a bool to denote when it's being disabled or not. This was attached to `isPhoneOpen` which wasn't reliablly working.  Its now done within `showPhone` and `hidePhone`.

The wait was changed from `1000` to `100` in the event the phone is opened before the wait is over. Makes it more smooth as there is no longer a small duration where the mouse moves the camera.

Removed the enable of control actions as it wasn't needed.