Fix only text twitter message
The option "allowNoMessage" not have logic, 
Explanation: It doesn't make much sense to force someone to have to put a text and an image and put it in a DEFAULT enabled configuration.
Also don't offer FEEDBACK to the user that they are doing something wrong, 80% of people are going to think "oh, it's a bug in the phone".
So I have removed this option from the configuration and the functionality is whether it has text OR it has an image. But if both are empty, the message cannot be sent.